+++ connect/api/v1/account/views.py
@@ -316,6 +316,31 @@ class MyUserProfileViewSet(
             print(e)
             return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR, data={"message": str(e)})
 
+    @action(
+        detail=True,
+        methods=["GET"],
+        url_name="user-company-info",
+        url_path="user-company-info",
+    )
+    def get_user_company_info(self, request, **kwargs):
+        user = self.request.user
+        authorizations = user.authorizations_user
+        if authorizations.count() == 1:
+            organization = authorizations.first().organization
+            first_user = organization.authorizations.order_by("created_at").first().user
+
+            if first_user.id != user.id:
+                data = dict(
+                    company_name=first_user.company_name,
+                    company_segment=first_user.company_segment,
+                    company_sector=first_user.company_sector,
+                    number_people=first_user.number_people,
+                    weni_helps=first_user.weni_helps
+                )
+                return Response(status=status.HTTP_200_OK, data=data)
+
+        return Response(status=status.HTTP_200_OK, data={})
+
 
 class SearchUserViewSet(mixins.ListModelMixin, GenericViewSet):  # pragma: no cover
     serializer_class = SearchUserSerializer