+++ connect/api/v1/tests/test_account.py
@@ -7,7 +7,9 @@ from rest_framework import status
 
 from connect.api.v1.account.views import MyUserProfileViewSet
 from connect.api.v1.tests.utils import create_user_and_token
+
 from unittest.mock import patch
+from connect.common.models import Organization, BillingPlan, OrganizationRole
 
 
 class ListMyProfileTestCase(TestCase):
@@ -153,3 +155,53 @@ class EmailVerifiedTestCase(TestCase):
     def test_okay(self, verify_email):
         response = self.request(self.user_token)
         self.assertEquals(response.status_code, status.HTTP_200_OK)
+
+
+class CompanyInfoTestCase(TestCase):
+    def setUp(self):
+        self.factory = RequestFactory()
+        self.owner, self.owner_token = create_user_and_token("owner")
+        self.user, self.user_token = create_user_and_token("user")
+
+        owner = self.owner
+
+        owner.company_name = "Test Company"
+        owner.company_segment = "Segment"
+        owner.company_sector = "Sector"
+        owner.number_people = 5
+        owner.weni_helps = ""
+        owner.save()
+
+        self.organization = Organization.objects.create(
+            name="test organization",
+            description="test organization",
+            inteligence_organization=1,
+            organization_billing__cycle=BillingPlan.BILLING_CYCLE_MONTHLY,
+            organization_billing__plan=BillingPlan.PLAN_TRIAL,
+        )
+        self.organization_authorization = self.organization.authorizations.create(
+            user=self.owner, role=OrganizationRole.ADMIN.value
+        )
+        self.organization_authorization2 = self.organization.authorizations.create(
+            user=self.user, role=OrganizationRole.ADMIN.value
+        )
+
+    def request(self, token):
+        authorization_header = {"HTTP_AUTHORIZATION": "Token {}".format(token.key)}
+        request = self.factory.get("/v1/account/user-company-info/", **authorization_header)
+        response = MyUserProfileViewSet.as_view({"get": "get_user_company_info"})(request)
+        response.render()
+        content_data = json.loads(response.content)
+        return (response, content_data)
+
+    def test_okay(self):
+        response, content_data = self.request(self.user_token)
+        expected_data = dict(
+            company_name=self.owner.company_name,
+            company_segment=self.owner.company_segment,
+            company_sector=self.owner.company_sector,
+            number_people=self.owner.number_people,
+            weni_helps=self.owner.weni_helps
+        )
+        self.assertEquals(content_data, expected_data)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)