+++ connect/template_projects/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
-from connect.template_projects.models import TemplateType, TemplateFeature, TemplateAI
+from connect.template_projects.models import TemplateType, TemplateFeature, TemplateSuggestion
 
 # Register your models here.
 admin.site.register(TemplateType)
 admin.site.register(TemplateFeature)
-admin.site.register(TemplateAI)
+admin.site.register(TemplateSuggestion)