+++ CHANGELOG.md
@@ -1,7 +1,30 @@
+# 3.4.0
+
+## *Add*
+  - New authorization endpoint on v2
+  - New unittests for billing tasks
+
+## *Fix*
+  - Flake8 issues
+
+# 3.3.0
+
+## *Delete*
+  - Remove TemplateAI api, models and unittests.
+
+# 3.2.1
+
+## *Add*
+  - register ready from template type
+  - internal endpoint to update project
+
+## *Fix*
+  - recent activity external user don't create a RecentActivity object
+
 # 3.2.0
 
 ## *Add*
-- Cursor pagination to v2 organizations and v2 projects
+  - Cursor pagination to v2 organizations and v2 projects
 
 # 3.1.0